chore: rename Types.ts to types.ts
