+++ tests/auth-util.test.ts
@@ -1,8 +1,8 @@
-import { TInternalConfig } from '../src/Types'
 import { fetchWithRefreshToken } from '../src/authentication'
 import { decodeJWT } from '../src/decodeJWT'
 import { FetchError } from '../src/errors'
 import { epochAtSecondsFromNow, epochTimeIsPast } from '../src/timeUtils'
+import { TInternalConfig } from '../src/types'
 
 const authConfig: TInternalConfig = {
   autoLogin: false,
