+++ tests/get_token.test.tsx
@@ -1,7 +1,7 @@
 import { render, waitFor } from '@testing-library/react'
 import React from 'react'
 import { AuthProvider } from '../src'
-import { TTokenResponse } from '../src/Types'
+import { TTokenResponse } from '../src/types'
 import { AuthConsumer, authConfig } from './test-utils'
 
 // @ts-ignore
